requirements.txt
Django>=4.2,<5.0
mysqlclient>=2.2.0
djangorestframework>=3.14.0
django-crispy-forms>=2.0
crispy-bootstrap5>=0.7
Pillow>=10.0.0
requests>=2.31.0
